+++ src/SWP/Bundle/FixturesBundle/DataFixtures/ORM/LoadContainersData.php
@@ -18,7 +18,6 @@ use Doctrine\Common\DataFixtures\FixtureInterface;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use SWP\Bundle\FixturesBundle\AbstractFixture;
 use Doctrine\Common\Persistence\ObjectManager;
-use SWP\Bundle\CoreBundle\Model\Container;
 
 class LoadContainersData extends AbstractFixture implements FixtureInterface, OrderedFixtureInterface
 {
@@ -27,14 +26,14 @@ class LoadContainersData extends AbstractFixture implements FixtureInterface, Or
      */
     public function load(ObjectManager $manager)
     {
-        $containerEntity = new Container();
-        $containerEntity->setName('container_name');
-        $containerEntity->setStyles('border: solid 1px red');
-        $containerEntity->setCssClass('col-md-12');
-        $manager->persist($containerEntity);
-        $manager->flush();
+        $env = $this->getEnvironment();
 
-        $this->addReference('container_name', $containerEntity);
+        $this->loadFixtures(
+            [
+                '@SWPFixturesBundle/Resources/fixtures/ORM/'.$env.'/Container.yml',
+            ],
+            $manager
+        );
     }
 
     public function getOrder()